simplify the problematic code
